+++ e2e/wdio/core/pages/date-time-picker.po.ts
@@ -42,7 +42,7 @@ export class DateTimePicker extends CoreBaseComponentPo {
     };
 
     dayInCalendarButtonByValue = (index: string): string => {
-        return `//span[contains(.,"${index}")]/ancestor::td[not (contains(@class, 'fd-calendar__item--other-month'))]`;
+        return `//span[contains(@id,"day-${index}-")]/ancestor::td[not (contains(@class, 'fd-calendar__item--other-month'))]`;
     };
 
     yearInCalendarByValue = (year: number): string => {