+++ libs/platform/src/lib/icon-tab-bar/directives/extra-button/extra-button.directive.ts
@@ -3,6 +3,8 @@ import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { ViewportRuler } from '@angular/cdk/overlay';
 
+const EXTRA_BUTTON_OFFSET_PX = 16;
+
 @Directive({
     selector: '[fdpExtraButton], [fdp-extra-button]'
 })
@@ -23,7 +25,7 @@ export class ExtraButtonDirective implements AfterViewInit, OnDestroy {
      * @description Offset to calculate correct position
      */
     @Input()
-    extraButtonOffset = 0;
+    addButtonOffset = false;
 
     /** @hidden */
     private _onDestroy$ = new Subject();
@@ -56,6 +58,13 @@ export class ExtraButtonDirective implements AfterViewInit, OnDestroy {
         const nativeEl = this._el.nativeElement;
         const parent = nativeEl.parentElement;
 
+        if (this.anchorIndexInsideParent === -1) {
+            this._el.nativeElement.style[this.isRtl ? 'right' : 'left'] = `${EXTRA_BUTTON_OFFSET_PX}px`;
+            return;
+        } else {
+            this._el.nativeElement.style.right = null;
+        }
+
         const anchor = parent.children[this.anchorIndexInsideParent];
         const computed = getComputedStyle(anchor);
 
@@ -65,9 +74,10 @@ export class ExtraButtonDirective implements AfterViewInit, OnDestroy {
 
         const margin = this.isRtl ? parseInt(computed.marginLeft, 10) : parseInt(computed.marginRight, 10);
 
+        const buttonOffset = this.addButtonOffset ? EXTRA_BUTTON_OFFSET_PX : 0;
         const myPositionWithOffset = this.isRtl
-            ? myPosition - margin - this.extraButtonOffset
-            : myPosition + margin + this.extraButtonOffset;
+            ? myPosition - margin - buttonOffset
+            : myPosition + margin + buttonOffset;
 
         this._el.nativeElement.style.left = `${myPositionWithOffset}px`;
     }