+++ libs/platform/src/lib/public_api.ts
@@ -26,3 +26,4 @@ export * from '@fundamental-ngx/platform/thumbnail';
 export * from '@fundamental-ngx/platform/upload-collection';
 export * from '@fundamental-ngx/platform/value-help-dialog';
 export * from '@fundamental-ngx/platform/wizard-generator';
+export * from '@fundamental-ngx/platform/icon-tab-bar';